重构t_alert_summary表

+++ cat-home/src/main/java/com/dianping/cat/core/mybatis/alert/summary/dao/AlertSummaryMapper.java
@@ -7,9 +7,9 @@ import org.apache.ibatis.annotations.Param;
 import com.dianping.cat.core.mybatis.alert.summary.dao.data.AlertSummaryDO;
 
 public interface AlertSummaryMapper {
-	int deleteByPrimaryKey(@Param("id") Integer id);
+	int deleteByPrimaryKey(@Param("id") Long id);
 
-	AlertSummaryDO findByPrimaryKey(@Param("id") Integer id);
+	AlertSummaryDO findByPrimaryKey(@Param("id") Long id);
 
 	int insert(AlertSummaryDO record);
